requirements-dev.txt
#
# Dependencies Required to Build `latexgit`
#

# minify_html is needed to minify html output.
minify_html == 0.16.4

# for converting the additional markdown files to HTML
markdown == 3.8

# for converting files to HTML
Pygments == 2.19.1

# for the PDF example
fpdf2 == 2.8.3
